JwtValidator is interface now Validators separated into the class files
