+++ security/jwt/src/main/java/io/helidon/security/jwt/JwtValidators.java
@@ -1,0 +1,41 @@
+/*
+ * Copyright (c) 2024 Oracle and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.helidon.security.jwt;
+
+import java.util.List;
+
+import io.helidon.common.Errors;
+
+/**
+ * Implementation of the JwtValidator interface.
+ */
+class JwtValidators implements JwtValidator {
+
+    private final List<ClaimValidator> claimValidators;
+
+    JwtValidators(Builder builder) {
+        claimValidators = List.copyOf(builder.claimValidators());
+    }
+
+    @Override
+    public Errors validate(Jwt jwt) {
+        Errors.Collector collector = Errors.collector();
+        claimValidators.forEach(claimValidator -> claimValidator.validate(jwt, collector, claimValidators));
+        return collector.collect();
+    }
+
+}
